changed: put episode videodatabase methods in separate file
